requirements.txt
django==2.2
djangorestframework==3.11.0
gunicorn==20.0.4
pandas==1.0.1
numpy==1.18.1
keras==2.3.1
matplotlib==3.2.0
pillow==7.0.0
scikit-learn==0.22.2
scipy==1.4.1
requests==2.23.0
tensorflow==1.13.1